Extract wallet settings views
